feat: go 1.24
