+++ .github/workflows/push.yaml
@@ -18,8 +18,8 @@ jobs:
     strategy:
       matrix:
         go:
-          - "1.22"
           - "1.23"
+          - "1.24"
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-go@v5
@@ -35,8 +35,8 @@ jobs:
     strategy:
       matrix:
         go:
-          - "1.22"
           - "1.23"
+          - "1.24"
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-go@v5
@@ -64,8 +64,8 @@ jobs:
       fail-fast: true
       matrix:
         go:
-          - "1.22"
           - "1.23"
+          - "1.24"
         environment:
           - development
           - staging
