+++ .github/workflows/server_regression.yaml
@@ -24,8 +24,8 @@ jobs:
     strategy:
       matrix:
         go:
-          - "1.22"
           - "1.23"
+          - "1.24"
         environment:
           - development
           - staging
